fix: Resolving file path changes after move
