+++ src/packages/hslString.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import ColorPicker from "./components/ColorPicker";
-import { ColorModel, ColorPickerBaseProps } from "./types";
-import { equalColorString } from "./utils/compare";
-import { hslStringToHsv, hsvToHslString } from "./utils/convert";
+import ColorPicker from "../components/ColorPicker";
+import { ColorModel, ColorPickerBaseProps } from "../types";
+import { equalColorString } from "../utils/compare";
+import { hslStringToHsv, hsvToHslString } from "../utils/convert";
 
 const colorModel: ColorModel<string> = {
   defaultColor: "hsl(0, 0%, 0%)",
