+++ src/packages/rgbString.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import ColorPicker from "./components/ColorPicker";
-import { ColorModel, ColorPickerBaseProps } from "./types";
-import { equalColorString } from "./utils/compare";
-import { rgbStringToHsv, hsvToRgbString } from "./utils/convert";
+import ColorPicker from "../components/ColorPicker";
+import { ColorModel, ColorPickerBaseProps } from "../types";
+import { equalColorString } from "../utils/compare";
+import { rgbStringToHsv, hsvToRgbString } from "../utils/convert";
 
 const colorModel: ColorModel<string> = {
   defaultColor: "rgb(0, 0, 0)",
